requirements.txt
git+git://github.com/alecthomas/injector.git@40fe9979138dc7c0b2eeeeec5aec9b88982a553e
pydns
xmpppy
requests
